+++ macros/src/log_baud_rate.rs
@@ -1,5 +1,5 @@
 use drone_config::Config;
-use drone_macros_core::compile_error;
+use drone_macros_core::parse_error;
 use proc_macro::TokenStream;
 use proc_macro2::Span;
 use quote::quote;
@@ -20,7 +20,7 @@ pub fn proc_macro(input: TokenStream) -> TokenStream {
     let Input {} = parse_macro_input!(input as Input);
     let config = match Config::read_from_cargo_manifest_dir() {
         Ok(config) => config,
-        Err(err) => compile_error!("{}: {}", drone_config::CONFIG_NAME, err),
+        Err(err) => parse_error!("{}: {}", drone_config::CONFIG_NAME, err),
     };
     let baud_rate = if let Some(log) = config.log {
         if let Some(swo) = log.swo {
@@ -28,13 +28,13 @@ pub fn proc_macro(input: TokenStream) -> TokenStream {
         } else if let Some(dso) = log.dso {
             dso.baud_rate
         } else {
-            compile_error!(
+            parse_error!(
                 "Missing one of `log.swo`, `log.dso` sections in `{}`",
                 drone_config::CONFIG_NAME
             );
         }
     } else {
-        compile_error!("Missing `log` section in `{}`", drone_config::CONFIG_NAME);
+        parse_error!("Missing `log` section in `{}`", drone_config::CONFIG_NAME);
     };
     let baud_rate = LitInt::new(&baud_rate.to_string(), Span::call_site());
     quote!(#baud_rate).into()